container-utils: Use the containerd API
Usage of the containerd API modifies some behavior of the
ContainerdClient. Instead of failing on the first API Call it will fail
when the ContainerdClient is created.
The default namespace for the ContainerdClient will remain "k8s.io" for
now. This assures backwards compatiblity.
Additionally all "sandbox" containers are filtered out.