update gadgets to use new interfaces
This adds the required implementations for GadgetDesc and Gadget to all gadgets. This commit tries to change as little as possible in the existing gadget code to keep backwards compatibility at the price of being a bit verbose or using not 100% fitting interfaces. We certainly need to clean this up once the architecture has been implemented for kubectl-gadget and gadgettracermgr as well.