Dockerfiles: Use busybox:glibc as base image for CO-RE image.
There are various concerns about alpine, particularly regarding the fact it uses
musl as libc which does not provide DNS over TCP [1].
Nonetheless, recent work was done to add this feature to musl hence to
alpine [2, 3].

So, this commit replaces alpine by busybox:glibc.
As a result, the binaries running inside this image will use glibc and not musl,
so there will be no problems regarding DNS.

This results in a decrease from 153MB to 141MB, i.e. 8.5%, for the CO-RE
container image.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1] https://medium.com/better-programming/why-i-will-never-use-alpine-linux-ever-again-a324fd0cbfd6
[2] https://gitlab.alpinelinux.org/alpine/tsc/-/issues/43#note_295619
[3] https://git.musl-libc.org/cgit/musl/commit/?id=51d4669fb97782f6a66606da852b5afd49a08001