gadgets: Add bpftrace initial integration
The initial integration of bpftrace in Inspektor Gadget consists of
executing the bpftrace binary on the gadget pod and streaming its
output to the client.

The current integration does not support filtering events by
namespace/pod/container nor enriching them with that information. This
is planned to be done in the second phase of this work.